fixed warnings that new version of compiler showed me

+++ common_src/common_utils/distribution.cpp
@@ -4,7 +4,7 @@
 #include "common_utils/utility.h"
 int base_seed = 1234;
 DistributionGenerator distibutionGenerator;
-#define RND ((float)rand()/RAND_MAX)
+#define RND ((double)rand()/RAND_MAX)
 Uniform *base = distibutionGenerator.get_uniform(0, 1, base_seed);
 UniformInt *basei = distibutionGenerator.get_uniform_int(0, INT_MAX, base_seed);
 Normal::Normal(double a, double sigma, int seed)
@@ -87,7 +87,7 @@ double *UniformInt::get_series(unsigned size)
 }
 double urand(double from, double to)
 {
-    return ((float)rand() / RAND_MAX) * (to - from) + from;
+    return ((double)rand() / RAND_MAX) * (to - from) + from;
 }
 double urandi(int from, int to)
 {
